By bringing it back to O(1) queries again

This speeds up the request time for students from 8s to 1.5s on my laptop and from 23 to 15s for staff users

Note that this is not the complete loading time for this page. my browser (firefox) needs another six seconds to finish processing/rendering the page.