@He3lixxx 2019-11-18
You removed the `stdout=StringIO()` parameter from the `management.call_command()` call - if was like this before.
```
management.call_command('refresh_results_cache', stdout=StringIO())
```
Did you _want_ to do this? I think I tested it and everything was fine - but with the current master, the stdout output is printed to the console again.