@karyon 2020-01-02
No, that was for debugging, i guess. I opened https://github.com/fsr-de/EvaP/pull/1399.